included anaCache for pvt samples + comments applied
